detector/requirements.txt
aio-pika
pydantic
pydantic-settings
psycopg2-binary
transformers
ultralytics
pillow
pycocotools
torchmetrics
torchvision
sqlalchemy
chromadb
supervision
boto3